Move the camera settings to the settings window
Move the camera sensitivity and movement speed to the settings window.
This makes the camera section of the left menu a little smaller.
